Extract HomeTagBar.tsx into its own component, so places besides EAHomeMainContent.tsx can reuse it
