+++ packages/lesswrong/lib/components.ts
@@ -1068,4 +1068,5 @@ importComponent("CookieBanner", () => require('../components/common/CookieBanner
 importComponent("CookieDialog", () => require('../components/common/CookieBanner/CookieDialog'));
 importComponent("CookiePolicy", () => require('../components/common/CookieBanner/CookiePolicy'));
 importComponent("CookieTable", () => require('../components/common/CookieBanner/CookieTable'));
+importComponent("HomeTagBar", () => require('../components/common/HomeTagBar'));
